Fix ILR lockup
This is a little bit of a hack. I changed the file ending of the svg to .twig so that `source()` call in _ilr-lockup.twig could find the template.

This is related to the UnionTwigLoader in union_organizer.

An alternate solution would have been to add this (or all) .svg files into the Twig environment so that source() could load them.