Prevent pointer events on floating labels in the form component
This is an attempt to fix an issue with Chrome and Safari when there are two inputs with the same id on the page. I know that this isn't supposed to happen, but it sometimes does, and this change prevents those browsers from jumping focus to the first form field with the same id.